Extract  component so it's easier later on to reuse in ai-assistant-chat-container
