+++ packages/ui-components/src/components/new-ai-chat-button.tsx
@@ -1,0 +1,35 @@
+import { t } from 'i18next';
+import { Plus } from 'lucide-react';
+import { useCallback } from 'react';
+import { Button } from '../ui/button';
+
+export const NewAiChatButton = ({
+  enableNewChat,
+  onNewChatClick,
+}: {
+  enableNewChat: boolean;
+  onNewChatClick: () => void;
+}) => {
+  const onClickHandler = useCallback(
+    (ev: React.MouseEvent<HTMLButtonElement>) => {
+      ev.stopPropagation();
+      onNewChatClick();
+    },
+    [onNewChatClick],
+  );
+  return (
+    <Button
+      variant="basic"
+      className="bg-gray-100 dark:bg-accent/10 bg-input rounded-xs mx-2"
+      size="xs"
+      onClick={onClickHandler}
+      disabled={!enableNewChat}
+      type="button"
+    >
+      <div className="flex items-center">
+        <Plus size={13} />
+        <span className="font-semibold text-xs ">{t('New chat')}</span>
+      </div>
+    </Button>
+  );
+};
